expenses modal fixed, expenses page deleted, form pages of bookings, documents and trips
