+++ src/components/PlanCard/PlanCard.jsx
@@ -5,7 +5,9 @@ import SavedPlanRow from "../SavedPlanRow/SavedPlanRow"
 import { useParams } from "react-router-dom"
 
 const PlanCard = () => {
-    console.log(useParams.id)
+    //console.log(useParams.id)
+    const { id } = useParams()
+    console.log(id)
     return (
 
         <div className="PlanCard">
