--- src/pages/Expenses/Expenses.jsx
@@ -1,12 +1,0 @@
-import HeaderExpenses from "../../components/Expenses/HeaderExpenses"
-
-const Expenses = () => {
-
-    return (
-        <div className="Exponses">
-            <HeaderExpenses />
-        </div>
-    )
-}
-
-export default Exponses
