+++ src/pages/TripDetail/TripDetail.css
